requirements.txt
boto3
pandas
requests
pyarrow
python-dotenv